backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pdfplumber==0.11.0
tiktoken==0.7.0
chromadb==1.0.20
numpy==1.26.4
scikit-learn==1.4.2
openai==1.99.9
python-dotenv==1.0.1
pydantic==1.10.22
pydantic_core==2.33.2
python-multipart==0.0.9

